+++ OpenORPG.Server/Game/Database/Models/ContentTemplates/SkillTemplate.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class SkillTemplate : IContentTemplate
     {
-        public SkillTemplate(string description, int id, string name, string garbage)
+        public SkillTemplate(string description, long id, string name, string garbage)
         {
             Description = description;
             Id = id;
@@ -15,7 +15,7 @@
 
         public string Description { get; set; }
         public string Garbage { get; set; }
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string VirtualCategory { get; set; }
     }
